Moved `safe_invoke()` in new `invoke.hpp` header.
