+++ libmamba/include/mamba/core/error_handling.hpp
@@ -5,7 +5,6 @@
 #include <stdexcept>
 #include <string>
 #include <vector>
-#include <functional>
 
 #include "tl/expected.hpp"
 
@@ -331,39 +330,6 @@ namespace mamba
     T&& extract(expected_ref_wrapper<T, E>&& exp)
     {
         return detail::extract_impl(std::move(exp));
-    }
-
-    template <typename Func, typename... Args>
-    auto safe_invoke(Func&& func, Args&&... args)
-        -> tl::expected<decltype(std::invoke(std::forward<Func>(func),
-                                             std::forward<Args>(args)...)),
-                        mamba_error>
-    {
-        auto call
-            = [&] { return std::invoke(std::forward<Func>(func), std::forward<Args>(args)...); };
-        using Result = decltype(call());
-        try
-        {
-            if constexpr (std::is_void<Result>::value)
-            {
-                call();
-                return {};
-            }
-            else
-            {
-                return call();
-            }
-        }
-        catch (const std::runtime_error& err)
-        {
-            return make_unexpected(std::string("callback invocation failed : ") + err.what(),
-                                   mamba_error_code::unknown);
-        }
-        catch (...)
-        {
-            return make_unexpected("callback invocation failed : unknown error",
-                                   mamba_error_code::unknown);
-        }
     }
 
 }
