+++ libmamba/src/core/execution.cpp
@@ -1,6 +1,6 @@
 #include "mamba/core/execution.hpp"
 
-#include "mamba/core/error_handling.hpp"
+#include "mamba/core/invoke.hpp"
 #include "mamba/core/output.hpp"
 
 namespace mamba
