+++ libmamba/tests/CMakeLists.txt
@@ -28,7 +28,7 @@ set(TEST_SRCS
     test_util.cpp
     test_env_lockfile.cpp
     test_execution.cpp
-    test_error_handling.cpp
+    test_invoke.cpp
     test_tasksync.cpp
 )
